requirements.txt
Django~=4.2.17
django-debug-toolbar~=4.4.6
djangorestframework~=3.15.2
drf-nested-routers~=0.94.1
django-filter~=24.3